vmm: invoke IRQ control capability for PPIs
When the VMM is running on a secondary core, the capDL-loader's IRQ
initialization will fail for PPIs, since each core has its own PPIs,
and the capDL-loader was running on core 0.

Invoking seL4_IRQControl_Get will tell the kernel to enable the PPI on
the VMM's core instead of the default boot core.

Change-Id: I8ef1ccc5a8fc5f52db9c0c78c8a8bd1bbd0dfd72