VM: Add IRQInit component
Giving a VM access to an irq control capability is not the best idea.
With that capability, the VMM can turn on and off every IRQ in the
entire system, which is a massive security vulnerability. Instead,
give that capability to a "3rd party" component: IRQInit. This component
provides an interface for the VM to request a PPI be turned on, getting
a capability to the handler in return.

Change-Id: Iea3d229ce7eaa2bb94a0516a00f2c74d4c6c373c